rename files to .ts, update to use es6 imports/exports, and build and watch scripts
