+++ package.json
@@ -38,12 +38,14 @@
     "jest-mock-console": "^0.4.2",
     "jest-mock-process": "^1.2.0",
     "tmp": "^0.1.0",
-    "typescript": "^3.4.5"
+    "typescript": "^4.2.3"
   },
   "scripts": {
+    "build": "tsc --project tsconfig.json",
+    "watch": "tsc --watch --project tsconfig.json",
     "test": "jest",
     "coverage": "jest --coverage --collectCoverageFrom=src/**/*.js --coveragePathIgnorePatterns=src/flow-to-ts.js",
-    "pretty-quick": "prettier src/*.js test/*.js playground/src/*.tsx babel-traverse/lib/**/*.js --write",
+    "pretty-quick": "prettier src/*.ts test/*.js playground/src/*.tsx babel-traverse/lib/**/*.js --write",
     "pretty-tests": "prettier ./test/fixtures/**/ts.js --parser typescript --write && prettier ./test/fixtures/**/flow.js --parser flow --write"
   },
   "husky": {
